synkron og asynkron prosessering
